backend/requirements.txt
rembg
fastapi
uvicorn
python-multipart
pillow
